feat: function to save files async
